Fixed up relations, modified layers

+++ src/main/java/com/revature/repository/ProfileRepository.java
@@ -1,6 +1,7 @@
 package com.revature.repository;
 
 import org.springframework.data.repository.CrudRepository;
+
 import com.revature.model.Profile;
 
 /**
@@ -9,7 +10,10 @@ import com.revature.model.Profile;
  * @author Adriana Long
  * @author Brian Abajah
  */
-public interface ProfileRepository extends CrudRepository <Profile, Long>{
+public interface ProfileRepository extends CrudRepository <Profile, Integer>{
     Profile findByLastName(String lastName);
     Profile findByFirstName(String firstName);
+    
+    //@Query("SELECT firstName, lastName FROM profile WHERE firstName + ' ' + lastName LIKE %:keyword%")
+    //List<Profile> search(@Param("keyword") String keyword);
 }
